Move functionHelper to js folder
