+++ app_project/jsx_files/report/project_report.js
@@ -1,6 +1,6 @@
 export { ProjectReport }
 
-import { functionHelper } from "./functionHelper.js"
+import { functionHelper } from "../functionHelper.js"
 
 class ProjectReport extends React.Component {
   constructor(props) {
