+++ app_project/jsx_files/report/search_report.js
@@ -1,6 +1,6 @@
 export { SearchReport }
 
-import { functionHelper } from "./functionHelper.js"
+import { functionHelper } from "../functionHelper.js"
 
 class SearchReport extends React.Component {
   constructor(props) {
