+++ public/javascripts/app_project/functionHelper.js
